# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Enterprise HR API Documentation",
        "description": "API Documentation for Team 11 Enterprise HR System, covering Employee, Attendance, and Approval modules.",
        "contact": {
            "email": "admin@enterprise.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/api/leaves": {
            "post": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "Submit a leave request",
                "operationId": "d0df8d57f277cce1e111041f7bdd45a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date",
                                    "reason",
                                    "type"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-03"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Family vacation"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "annual"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leave request submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LeaveRequest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or insufficient leave balance"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/leaves/my-requests": {
            "get": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "View leave history and status by authenticated employee",
                "operationId": "0cb85491bc9cfb2413318ac3c87fd6eb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "leaves_balances": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LeaveRequest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/leaves/subordinates": {
            "get": {
                "tags": [
                    "Leave Request"
                ],
                "summary": "View all leaves from subordinates only",
                "operationId": "376b6bf5228fa703f9a06f4d23f6d3d7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LeaveRequest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LeaveRequest": {
                "title": "Leave Request Model",
                "description": "Employee leave request",
                "required": [
                    "id",
                    "employee_id",
                    "start_date",
                    "end_date",
                    "reason",
                    "type",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employee_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-10"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-12"
                    },
                    "reason": {
                        "type": "string",
                        "example": "Vacation"
                    },
                    "type": {
                        "type": "string",
                        "example": "annual"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}